pyproject.toml
[project]
name = "jobsherpa"
version = "0.1.0"
description = "An AI agent for HPC systems."
authors = [{ name = "Your Name", email = "you@example.com" }]
dependencies = [
    "typer",
    "PyYAML",
    "farm-haystack",
    "setuptools",
    "Jinja2",
    "pydantic",
    "ruamel.yaml",
]

[project.scripts]
jobsherpa = "jobsherpa.cli.main:run_app"

[project.optional-dependencies]
test = [
    "pytest",
    "freezegun",
    "pytest-cov",
    "pytest-mock",
]
inference = ["farm-haystack[inference]"]

[tool.setuptools.packages.find]
where = ["."]
include = ["jobsherpa*"]
exclude = ["tests*"]
